@@ -14,16 +14,16 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.12.0+2
+  http: ^1.1.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.5
 
 dev_dependencies:
   flutter_test:
@@ -43,12 +43,10 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
-   - 1.jpg
-   - 2.jpg
-   - 6.jpg
-   - images/3.jpg
-   - images/4.jpg
-  #
+   - images/1.jpg
+   - images/2.jpg
+   - images/6.jpg
+   
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
   fonts:
