@@ -14,18 +14,18 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.12.0+2
-  firebase_core: ^0.4.0+6  # main dependency for Firebase Core
-  cloud_firestore: ^0.12.7 # dependency for Firebase Cloud Storage
+  http: ^1.1.0
+  firebase_core: ^2.15.0  # main dependency for Firebase Core
+  cloud_firestore: ^4.8.4 # dependency for Firebase Cloud Storage
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.5
 
 dev_dependencies:
   flutter_test:
