@@ -24,7 +24,7 @@ class SnackBarAlert extends StatelessWidget {
                     onPressed: () {},
                   ),
                 );
-                Scaffold.of(context).showSnackBar(sb);
+                ScaffoldMessenger.of(context).showSnackBar(sb);
               }))
           .toList(),
     );
@@ -37,7 +37,7 @@ class SnackBarDemo extends StatefulWidget {
 }
 
 class _SnackBarDemoState extends State<SnackBarDemo> {
-  List<dynamic> _peopleList = List<dynamic>();
+  List<dynamic> _peopleList = <dynamic>[];
 
   void _getPeople() async {
     String peopleString = await DefaultAssetBundle.of(context)
@@ -108,10 +108,8 @@ class PersonCard extends StatelessWidget {
                   ),
                 ),
                 child: Text(fullName,
-                    style: TextStyle(
-                      color: Colors.white,
-                      fontStyle: Theme.of(context).textTheme.title.fontStyle,
-                    )),
+                  style: Theme.of(context).textTheme.titleMedium?.copyWith(color: Colors.white),
+                ),
               ),
               Spacer(),
               Container(
