@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 class ColorCircle extends StatelessWidget {
   final Color color;
   final double radius;
-  ColorCircle({@required this.color, this.radius=50});
+  ColorCircle({required this.color, this.radius=50});
 
   @override
   Widget build(BuildContext context) {
