@@ -2,11 +2,11 @@ import 'package:flutter/material.dart';
 
 class ColorValueChanger extends StatefulWidget {
   final String property;
-  final Function onChanged;
+  final Function? onChanged;
   final int value;
   // Value passed in from its host
   ColorValueChanger(
-      {Key key,
+      {Key? key,
       this.property = "property",
       this.value = 0,
       this.onChanged})
@@ -16,18 +16,18 @@ class ColorValueChanger extends StatefulWidget {
 }
 
 class _ColorValueChangerState extends State<ColorValueChanger> {
-  int _value;
+  int? _value;
   @override
   Widget build(BuildContext context) {
     _value = widget.value;
     return Container(
       child: Column(
         children: <Widget>[
-          Text(widget.property, style: Theme.of(context).textTheme.subtitle),
+          Text(widget.property, style: Theme.of(context).textTheme.titleMedium),
           Slider(
             min: 0,
             max: 255,
-            value: _value?.toDouble(),
+            value: _value!.toDouble(),
             label: widget.property,
             onChanged: _onChanged,
           )
@@ -40,6 +40,6 @@ class _ColorValueChangerState extends State<ColorValueChanger> {
     setState(() {
       this._value = value.round();
     });
-    widget.onChanged(widget.property, value.round());
+    widget.onChanged!(widget.property, value.round());
   }
 }
