@@ -14,7 +14,7 @@ class BoxModelDemo extends StatefulWidget {
 }
 
 class _BoxModelDemoState extends State<BoxModelDemo> {
-  List<dynamic> _peopleList = List<dynamic>();
+  List<dynamic> _peopleList = <dynamic>[];
 
   void _getPeople() async {
     String peopleString = await DefaultAssetBundle.of(context)
