@@ -35,7 +35,7 @@ class ContainerDemo extends StatefulWidget {
 }
 
 class _ContainerDemoState extends State<ContainerDemo> {
-  List<dynamic> _peopleList = List<dynamic>();
+  List<dynamic> _peopleList = <dynamic>[];
 
   void _getPeople() async {
     String peopleString = await DefaultAssetBundle.of(context)
