@@ -7,7 +7,7 @@ class Widget1 extends StatelessWidget {
       child: Container(
           child:  Text(
         "I'm widget 1",
-        style: Theme.of(context).textTheme.display1,
+        style: Theme.of(context).textTheme.displayLarge,
       )),
     );
   }
