@@ -7,7 +7,7 @@ class Widget3 extends StatelessWidget {
       child: Container(
         child: Text(
           "I'm widget 3",
-          style: Theme.of(context).textTheme.display1,
+          style: Theme.of(context).textTheme.displayLarge,
         ),
       ),
     );
