@@ -13,7 +13,7 @@ class WidgetWithDrawer2 extends StatelessWidget {
         child: Container(
             child: Text(
           "I'm widget 2",
-          style: Theme.of(context).textTheme.display1,
+          style: Theme.of(context).textTheme.displayLarge,
         )),
       ),
     );
